// client/src/app/shared/flattened-small-provider.model.ts
export class FlattenedSmallProvider {
  entityTypeCode: string;
  entityTypeDisplayName: string;
  enumerationDate;
  firstLinePracticeLocationAddress: string;
  firstName: string;
  genderCodeid: string;
  id: number;
  lastName: string;
  lastUpdateDate: string;
  middleName: string;
  npi: string;
  organizationName: string;
  practiceLocationAddressCityName: string;
  practiceLocationAddressCountryCode: string;
  practiceLocationAddressFaxNumber: string;
  practiceLocationAddressPostalCode: string;
  practiceLocationAddressStateName: string;
  practiceLocationAddressTelephoneNumber: string;
  secondLinePracticeLocationAddress: string;
  system: string;
}